third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-srcset-react-to-media-changes-when-viewport-downsized-expected.txt

PASS img.currentSrc.indexOf("image-set-4x") is not -1
PASS successfullyParsed is true

TEST COMPLETE
